requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jinja2
aiofiles